+++ noodles-gtf/src/lib.rs
@@ -5,9 +5,11 @@
 pub mod io;
 pub mod line;
 pub mod record;
-mod writer;
 
-pub use self::{line::Line, record::Record, writer::Writer};
+pub use self::{line::Line, record::Record};
 
 #[deprecated(since = "0.32.0", note = "Use `noodles_gtf::io::Reader` instead.")]
 pub use self::io::Reader;
+
+#[deprecated(since = "0.32.0", note = "Use `noodles_gtf::io::Writer` instead.")]
+pub use self::io::Writer;